requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.8
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
mccabe==0.6.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0